app/requirements.txt
# Backend
requests
django

# Infrastructure)
psycopg2-binary
gunicorn
raven

# API
djangorestframework
django-cors-headers

# Static files
whitenoise

# Admin
django-extensions
ipython

# File storage
boto3
django-storages
Pillow

# Linting
isort
black
flake8

# Testing
faker
pytest-django
pytest-factoryboy
pytest-sugar
responses
freezegun
